fix(camel-redis): fix all 7 failing Redis integration tests
- Fix BLPOP timeout type: pass u64 instead of f64 (Redis rejects floats)
- Fix log::debug! -> tracing::debug! in producer (wrong logging crate)
- Rewrite redis_test.rs: use CamelRedis.Value headers instead of map_body,
  correct connection string format, add route IDs, error handler for
  diagnostics, and remove needless &format!() borrows

+++ crates/components/camel-redis/src/consumer.rs
@@ -260,7 +260,7 @@ async fn run_queue_consumer(
             result = async {
                 let cmd = redis::cmd("BLPOP")
                     .arg(&key)
-                    .arg(timeout as f64)
+                    .arg(timeout)
                     .to_owned();
                 cmd.query_async::<Option<(String, String)>>(&mut conn).await
             } =>
